Fixed Recipients state showing null Organized alias folder/files
